python-mcp/pyproject.toml
[project]
name = "papr-memory-demo"
version = "0.1.0"
description = "Demo project for Papr Memory API integration"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.26.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
debug-server = "debugpy:run" 